refactor: выделила в header меню из виджетов
