+++ src/layouts/Header/ui/MobileMenu/MobileMenu.props.ts
