--- src/widgets/index.ts
@@ -1,2 +1,0 @@
-export * from './Menu';
-
